refactor(agents): de-duplicate the agents/tests doubles and split the file at the size cap
Preparation for wiring agents/tests/ into CI. Behaviour-neutral: the suite
reports the same 8 failed / 414 passed before and after, by the same names.

WHY THE DOUBLES ARE CONSOLIDATED. test_inject_memory_context.py and
test_inject_memory_context_gates.py each carried their own near-copy of
_FakeEpisode / _FakeNote — the gates file's header even said "mirrors
test_inject_memory_context.py". These fakes stand in for real persistence
rows, so every field production grows is a field they must grow too, and that
debt was owed twice and paid neither time: both copies had fallen behind
Episode and Note in the same three ways. Two copies is why one production
change breaks two files, so they move into _memory_context_helpers.py, per the
existing _scheduled_wakes_wiring_helpers.py convention. The gates file's
leaner variant is dropped in favour of the superset; it differed only by
lacking make_mixin's `rel` parameter and FakeNote.id, neither of which
changes how its tests behave.

WHY THE TICK FILE IS SPLIT. test_persona_tick_shortcircuit.py sat at EXACTLY
500 lines — the code size cap enforced by scripts/checks/file_size.py --strict
in the pre-commit hook and in CI. At the cap, every subsequent fix has to pay
for itself by deleting existing rationale, which is how comments explaining
past defects get lost. The DB-failure contract is a distinct concern from the
short-circuit cases and is the natural seam, so it moves to
test_persona_tick_db_failure.py with the shared config + builders extracted to
_persona_tick_helpers.py. Both halves now sit well under the cap (352 and 89),
so the repair that follows this commit does not have to buy room.

Both extractions are pure moves: no test body is edited, no assertion changes,
and the definitions are the ones the files already carried.
