+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-08-29  
-> **2006** tracked files
+> **2009** tracked files
 
 ## Summary
 
@@ -10,7 +10,7 @@
 | `internal/` | 515 |
 | `docs/` | 454 |
 | `tests/` | 429 |
-| `agents/` | 282 |
+| `agents/` | 285 |
 | `web/` | 78 |
 | `prompts/` | 48 |
 | `cli/` | 42 |
@@ -28,7 +28,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 738 |
+| `.py` | 741 |
 | `.go` | 533 |
 | `.md` | 519 |
 | `.yaml` | 57 |
@@ -323,6 +323,8 @@
 │   │   └── rendering.py
 │   ├── tests/
 │   │   ├── __init__.py
+│   │   ├── _memory_context_helpers.py
+│   │   ├── _persona_tick_helpers.py
 │   │   ├── _scheduled_wakes_wiring_helpers.py
 │   │   ├── conftest.py
 │   │   ├── test_action_loop_channel_lease.py
@@ -356,6 +358,7 @@
 │   │   ├── test_observability_metrics.py
 │   │   ├── test_observability_spans.py
 │   │   ├── test_observability_tracing.py
+│   │   ├── test_persona_tick_db_failure.py
 │   │   ├── test_persona_tick_shortcircuit.py
 │   │   ├── test_polish_pr6.py
 │   │   ├── test_scheduled_wakes_cache.py
