+++ agents/tests/test_persona_tick_shortcircuit.py
@@ -20,99 +20,25 @@ Module pin (RFC 0017 PR plan open-at-plan-time resolution):
 
 from __future__ import annotations
 
-import copy
 import logging
-from typing import Any
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from agents.llm_client import LLMClient, LLMResponse, StopReason, Usage
-from agents.persona import create_persona_agent
-from agents.persona_runtime import _LLMPersonaAgent
-from agents.persona_runtime.memory_context import MemoryInjectionResult
 from agents.persona_types import ActionType, AgentAction, AgentEvent, EventType
+from agents.tests._persona_tick_helpers import (
+    make_agent as _make_agent,
+)
+from agents.tests._persona_tick_helpers import (
+    make_client as _make_client,
+)
+from agents.tests._persona_tick_helpers import (
+    nonzero_injection as _nonzero_injection,
+)
+from agents.tests._persona_tick_helpers import (
+    zero_injection as _zero_injection,
+)
 from agents.tick import TickScheduler
-
-# ─── Helpers ──────────────────────────────────────────────────────────────────
-
-
-_PERSONA_CONFIG: dict[str, Any] = {
-    "id": "test-agent",
-    "type": "persona",
-    "name": "Test Agent",
-    "role": "Testing",
-    "model": "test-model",
-    "temperature": 0.7,
-    "max_llm_calls": 10,
-    "max_tokens": 4096,
-    "persona": {
-        "background": "Test background.",
-        "behavior": {},
-    },
-    "permissions": {
-        "memory": {"read": True, "write": True},
-    },
-    "memory": {
-        "db_path": ":memory:",
-        "notes": {"max_notes": 100, "auto_reflect_after": 5},
-    },
-}
-
-_DO_NOTHING_RESPONSE = LLMResponse(
-    text='[{"action_type": "do_nothing", "payload": {}}]',
-    stop_reason=StopReason.END_TURN,
-    usage=Usage(input_tokens=10, output_tokens=5),
-)
-
-_SUBSTANTIVE_RESPONSE = LLMResponse(
-    text='[{"action_type": "complete_task", "payload": {"result": "done"}}]',
-    stop_reason=StopReason.END_TURN,
-    usage=Usage(input_tokens=10, output_tokens=10),
-)
-
-
-def _make_client(response: LLMResponse = _SUBSTANTIVE_RESPONSE) -> LLMClient:
-    """Return a mock LLMClient that returns the given response."""
-    mock_provider = AsyncMock()
-    mock_provider.create_message = AsyncMock(return_value=response)
-    mock_provider.format_tool_definitions = MagicMock(return_value=[])
-    mock_provider.append_tool_round = MagicMock(return_value=[])
-    return LLMClient(mock_provider)
-
-
-async def _make_agent(
-    *,
-    goal_progress: dict[str, float] | None = None,
-    recent_context: list[str] | None = None,
-    client: LLMClient | None = None,
-) -> _LLMPersonaAgent:
-    """Create an initialized _LLMPersonaAgent with optional state preset."""
-    agent = create_persona_agent(
-        agent_id="test-agent",
-        # Deep-copy so per-test mutations of nested dicts (``persona``,
-        # ``permissions``, ``memory``) cannot leak across tests via the
-        # shared module-level _PERSONA_CONFIG reference.
-        config=copy.deepcopy(_PERSONA_CONFIG),
-        llm_client=client or _make_client(),
-    )
-    await agent.initialize_memory()
-    if goal_progress is not None:
-        agent._state.goal_progress = goal_progress
-    if recent_context is not None:
-        agent._state.recent_context = recent_context
-    return agent
-
-
-def _zero_injection() -> MemoryInjectionResult:
-    """Return a MemoryInjectionResult with zero admitted tokens."""
-    return MemoryInjectionResult(memory_admitted_tokens=0)
-
-
-def _nonzero_injection(tokens: int = 200) -> MemoryInjectionResult:
-    """Return a MemoryInjectionResult with non-zero admitted tokens."""
-    return MemoryInjectionResult(memory_admitted_tokens=tokens)
-
 
 # ─── Core short-circuit cases ─────────────────────────────────────────────────
 
@@ -424,77 +350,3 @@ class TestAccessors:
     async def test_has_pending_turn_true_when_nonempty(self) -> None:
         agent = await _make_agent(recent_context=["some context"])
         assert agent._has_pending_turn() is True
-
-
-# ─── DB-failure path (tier-exception swallowing) ──────────────────────────────
-
-
-class TestDBFailurePath:
-    """All memory-tier lookups raise → admitted=0 → tick suppressed.
-
-    This class closes the coverage gap flagged in the PR #149 deep review
-    (Nice-to-have #1).  The other tests in this file patch
-    ``_inject_memory_context`` directly to return a zero-token result,
-    which proves the *guard* fires on a zero result but does not prove
-    the *upstream swallowing path* actually produces a zero result when
-    every tier lookup raises.
-
-    RFC 0017 §F documents — and ``_on_event_inner`` explicitly comments
-    on — the design decision that DB failure is intentionally indistinguishable
-    from a genuine empty-context tick: both lead to suppression, with
-    per-tier ``logger.warning`` calls in ``_inject_memory_context`` as the
-    operator-visible signal.  This test pins that contract end-to-end so a
-    future refactor that, e.g., re-raises tier exceptions instead of
-    swallowing them would fail loudly here.
-    """
-
-    @pytest.mark.asyncio
-    async def test_all_tier_lookups_raising_suppresses_tick(
-        self, caplog: pytest.LogCaptureFixture
-    ) -> None:
-        """Episodic + notes recall both raise → DO_NOTHING, no LLM call, warnings logged.
-
-        TICK events have no ``sender_id`` so ``_inject_memory_context`` never
-        invokes the relationship tier (see the ``if sender_id:`` guard in
-        ``memory_context.py``).  Patching the two tiers that *are* exercised
-        on a TICK is sufficient to drive ``memory_admitted_tokens`` to zero
-        through the swallowing path.
-        """
-        client = _make_client()
-        agent = await _make_agent(client=client)
-
-        with (
-            patch.object(
-                agent._episodic_memory,
-                "recall",
-                side_effect=OSError("DB unavailable"),
-            ),
-            patch.object(
-                agent._episodic_memory,
-                "recall_notes",
-                side_effect=OSError("DB unavailable"),
-            ),
-            caplog.at_level(
-                logging.WARNING,
-                logger="agents.persona_runtime.memory_context",
-            ),
-        ):
-            actions = await agent.on_event(AgentEvent(event_type=EventType.TICK))
-
-        # Suppression: all exercised tiers raised → admitted=0 → DO_NOTHING.
-        assert len(actions) == 1
-        assert actions[0].action_type == ActionType.DO_NOTHING
-        client._provider.create_message.assert_not_called()  # type: ignore[attr-defined]
-
-        # Operator-visible signal: each failed tier emitted a warning.
-        # Use ``getMessage()`` rather than ``.message`` to resolve the
-        # ``%s`` placeholders in the format string.
-        failure_warnings = [
-            r for r in caplog.records
-            if r.levelno == logging.WARNING and "failed, skipping" in r.getMessage()
-        ]
-        assert len(failure_warnings) >= 2, (
-            f"Expected at least 2 tier-failure warnings (episodic + notes), "
-            f"got {len(failure_warnings)}: "
-            f"{[r.getMessage() for r in failure_warnings]}"
-        )
